+++ apps/astro-cloudflare-playground/lucid.config.ts
@@ -55,7 +55,7 @@ export default configureLucid({
 			pagesPlugin({
 				collections: [
 					{
-						collection: PageCollection.key,
+						key: PageCollection.key,
 						localized: true,
 						prefix: {
 							en: "en",