+++ apps/cloudflare-playground/lucid.config.ts
@@ -29,7 +29,7 @@ export default configureLucid({
 			pagesPlugin({
 				collections: [
 					{
-						collection: PageCollection.key,
+						key: PageCollection.key,
 					},
 				],
 			}),