+++ packages/plugin-pages/src/services/hooks/after-fetch-handler.test.ts
@@ -24,7 +24,7 @@ const context = {
 const options = {
 	collections: [
 		{
-			collection: "pages",
+			key: "pages",
 			localized: false,
 			ui: {
 				fullSlug: true,