+++ templates/cloudflare-deploy/lucid.config.ts
@@ -22,7 +22,7 @@ export default configureLucid({
 			pagesPlugin({
 				collections: [
 					{
-						collection: PageCollection.key,
+						key: PageCollection.key,
 					},
 				],
 			}),